Refactored ingestion logic to utilize API functions for database operations, improving code organization and readability.
